requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
easyocr==1.7.1
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pdf2image==1.16.3
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
requests==2.31.0
